Added cover and background
