+++ src/features/account/screens/account.screen.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+import { AccountBackground, AccountCover } from "../components/account.styles";
+
+export const AccountScreen = () => {
+  return (
+    <AccountBackground>
+      <AccountCover />
+    </AccountBackground>
+  );
+};
