+++ src/features/account/screens/login.screen.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const LoginScreen = () => {
+  return (<></>);
+};
