+++ src/features/account/screens/register.screen.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const RegisterScreen = () => {
+  return (<></>);
+}
